+++ prr-core/src/prr/clients/GoldClient.java
@@ -26,15 +26,15 @@ public class GoldClient extends ClientLevel {
         }
     }
 
-    public double getTextCost(int length) {
+    public long getTextCost(int length) {
         return _client.getTariff().getTextCostGold(length);
     }
 
-    public double getVoiceCost(int duration) {
+    public long getVoiceCost(int duration) {
         return _client.getTariff().getVoiceCostGold(duration);
     }
 
-    public double getVideoCost(int duration) {
+    public long getVideoCost(int duration) {
         return _client.getTariff().getVideoCostGold(duration);
     }
